<==== Mercurial data ====>
Branch: hls

+++ client/src/ui/workbench/workbench_controller.cpp
@@ -1056,7 +1056,7 @@ void QnWorkbenchController::at_zoomRectCreated(QnMediaResourceWidget *widget, co
 void QnWorkbenchController::at_zoomTargetChanged(QnMediaResourceWidget *widget, const QRectF &zoomRect, QnMediaResourceWidget *zoomTargetWidget) {
     QnLayoutItemData data = widget->item()->data();
     data.uuid = QUuid::createUuid();
-    data.resource.id = zoomTargetWidget->resource()->toResource()->getId().toInt();
+    data.resource.id = zoomTargetWidget->resource()->toResource()->getId();
     data.resource.path = zoomTargetWidget->resource()->toResource()->getUniqueId();
     data.zoomTargetUuid = zoomTargetWidget->item()->uuid();
     data.rotation = zoomTargetWidget->item()->rotation();
